Fix toon translator #4165

+++ plugins/mtoa/translators/shader/ShaderTranslators.cpp
@@ -1864,9 +1864,9 @@ void CProjectionTranslator::Export(AtNode* shader)
       MEulerRotation rot = m.eulerRotation();
       MVector trans = m.getTranslation(MSpace::kTransform);
 
-      AiNodeSetVec(shader, "scale", 0.5f/scale[0], 0.5f/scale[1], 0.5f/scale[2]);
-      AiNodeSetVec(shader, "offset", trans.x, trans.y, trans.z);
-      AiNodeSetVec(shader, "rotate", rot.x, rot.y, rot.z);
+      AiNodeSetVec(shader, "scale", float(0.5f/scale[0]), float(0.5f/scale[1]), float(0.5f/scale[2]));
+      AiNodeSetVec(shader, "offset", (float)trans.x, (float)trans.y, (float)trans.z);
+      AiNodeSetVec(shader, "rotate", (float)rot.x, (float)rot.y, (float)rot.z);
 
    } else if (AiNodeIs(shader, AtString("camera_projection")))
    {
